# swagger/swagger.json
{
  "swagger": "2.0",
  "host": "0.0.0.0:40000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "info": {
    "title": "Book Stall",
    "version": "1.0",
    "description": "Book Stall Api Documentation",
    "contact": {
      "name": "admin",
      "email": "admin@admin.com"
    }
  },
  "tags": [
    {
      "name": "Book Stall",
      "description": "Api interface."
    }
  ],
  "paths": {
    "/book": {
      "get": {
        "summary": "Get Existing Book List",
        "operationId": "getBook",
        "description": "Get Existing Book List",
        "parameters": [
          {
            "type": "number",
            "description": "Page number starting from 1",
            "minimum": 0,
            "name": "page_no",
            "in": "query"
          },
          {
            "type": "number",
            "description": "Page size",
            "minimum": 1,
            "name": "page_size",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Column name to sort",
            "enum": [
              "city",
              "start_date",
              "end_date",
              "price",
              "status",
              "color"
            ],
            "name": "sort_by",
            "in": "query"
          },
          {
            "type": "number",
            "description": "Order to sort",
            "enum": [
              1,
              -1
            ],
            "name": "sort_order",
            "in": "query"
          },
          {
            "type": "string",
            "name": "filter",
            "in": "query"
          }
        ],
        "tags": [
          "book"
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Model1"
            },
            "description": "Successful"
          },
          "201": {
            "schema": {
              "$ref": "#/definitions/Model1"
            },
            "description": "Created"
          },
          "400": {
            "schema": {
              "$ref": "#/definitions/Model1"
            },
            "description": "Bad Request"
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/Model1"
            },
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "Add New Book",
        "operationId": "postBook",
        "description": "Add New Book",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Model2"
            }
          }
        ],
        "tags": [
          "book"
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Model1"
            },
            "description": "Successful"
          },
          "201": {
            "schema": {
              "$ref": "#/definitions/Model1"
            },
            "description": "Created"
          },
          "400": {
            "schema": {
              "$ref": "#/definitions/Model1"
            },
            "description": "Bad Request"
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/Model1"
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/book/{id}": {
      "get": {
        "summary": "Get Existing Book",
        "operationId": "getBookId",
        "description": "Get Existing Book",
        "parameters": [
          {
            "type": "string",
            "description": "Book id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "tags": [
          "book"
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Model1"
            },
            "description": "Successful"
          },
          "201": {
            "schema": {
              "$ref": "#/definitions/Model1"
            },
            "description": "Created"
          },
          "400": {
            "schema": {
              "$ref": "#/definitions/Model1"
            },
            "description": "Bad Request"
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/Model1"
            },
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "summary": "Update existing Book",
        "operationId": "putBookId",
        "description": "Update existing Book",
        "parameters": [
          {
            "type": "string",
            "description": "Book id",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Model3"
            }
          }
        ],
        "tags": [
          "book"
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Model1"
            },
            "description": "Successful"
          },
          "201": {
            "schema": {
              "$ref": "#/definitions/Model1"
            },
            "description": "Created"
          },
          "400": {
            "schema": {
              "$ref": "#/definitions/Model1"
            },
            "description": "Bad Request"
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/Model1"
            },
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "summary": "Delete Existing Book",
        "operationId": "deleteBookId",
        "description": "Delete Existing Book",
        "parameters": [
          {
            "type": "string",
            "description": "Book id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "tags": [
          "book"
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Model1"
            },
            "description": "Successful"
          },
          "201": {
            "schema": {
              "$ref": "#/definitions/Model1"
            },
            "description": "Created"
          },
          "400": {
            "schema": {
              "$ref": "#/definitions/Model1"
            },
            "description": "Bad Request"
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/Model1"
            },
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
    "Model1": {
      "type": "object",
      "properties": {
        "is_success": {
          "type": "boolean"
        },
        "result": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "status_code": {
          "type": "number"
        }
      },
      "required": [
        "message",
        "status_code"
      ]
    },
    "Model2": {
      "type": "object",
      "properties": {
        "author": {
          "type": "string",
          "description": "Author Name",
          "example": "John Doe"
        },
        "title": {
          "type": "string",
          "description": "Book Title",
          "example": "John Biography"
        },
        "isbn": {
          "type": "string",
          "description": "ISBN of book",
          "example": "978-3-16-148410-0"
        },
        "release_date": {
          "type": "string",
          "format": "date",
          "description": "Release Date",
          "example": "2020-10-16"
        }
      },
      "required": [
        "author",
        "title",
        "isbn",
        "release_date"
      ]
    },
    "Model3": {
      "type": "object",
      "properties": {
        "author": {
          "type": "string",
          "description": "Author Name",
          "example": "John Doe"
        },
        "title": {
          "type": "string",
          "description": "Book Title",
          "example": "John Biography"
        },
        "isbn": {
          "type": "string",
          "description": "ISBN of book",
          "example": "978-3-16-148410-0"
        },
        "release_date": {
          "type": "string",
          "format": "date",
          "description": "Release Date",
          "example": "2020-10-16"
        }
      }
    }
  }
}